backend/scraper/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
pytesseract>=0.3.10
easyocr>=1.7.0
opencv-python-headless>=4.9.0
numpy>=1.24.0
Pillow>=10.2.0
supabase>=2.3.0
python-dotenv>=1.0.0
